// src/main/ios-tools/types.ts
export interface IOSResult<T> {
  success: boolean;
  data?: T;
  error?: string;
  errorCode?: string;
}